requirements.txt
asgiref==3.2.3
asn1crypto==1.2.0
attrs==19.3.0
Automat==0.8.0
bleach==3.1.0
boto3==1.10.38
botocore==1.13.38
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coincurve==13.0.0
constantly==15.1.00
dj-database-url==0.5.0
Django==3.0
docutils==0.15.2
ecdsa==0.14.1
eciespy==0.3.2
eth-hash==0.2.0
eth-keys==0.3.1
eth-typing==2.2.1
eth-utils==1.8.4
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jmespath==0.9.4
keyring==20.0.0
numpy==1.17.4
pkginfo==1.5.0.1
psycopg2-binary==2.8.4
pycparser==2.19
pycryptodome==3.9.4
Pygments==2.5.2
PyHamcrest==1.9.0
python-dateutil==2.8.0
pytz==2019.3
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
tqdm==4.40.2
urllib3==1.25.7
virtualenv==16.7.8
webencodings==0.5.1
whitenoise==5.0.1
zope.interface==4.7.1
